HW_4/app/requirements.txt
flask==2.2.2
gunicorn==20.1.0
Werkzeug==2.2.2
psycopg2-binary==2.9.9